Extract parser query grammar
